worker_node/requirements.txt
boto3
botocore
PyYAML
whisper
torch

